Recover outside-hull bodies and tighten comparison mismatches
Recover the eight remaining outside-hull Wizardry functions:
SettlePositionToGround00420C30, stSurface2D::render/setScale,
GetWorldCursorNode0048ED10, AttachWorldCursorNode0048ED30,
SetWorldCursorNodeName0048F110, UpdateCameraPathStateByName,
GetCharacterProjectilePosition004A57B0 and
W8CameraShakeEffect::Evaluate004AE4E0.

Then a comparison-quality batch across paired functions:

- LevelData flag getters (4/6/8/9) return unsigned char per retail
  byte-return codegen; ClampCameraToBounds last bound check spelled
  as retail's camera.z <= maximum->z structure.
- W8Navigator::SetPitchRollEnabled now exact: pitch/roll fields are
  char, storing raw parameter bytes as retail does.
- W8PropRepresentation::ToggleAnimation, stMeshModel::getBoundingBox
  and getBoundingSphere are exact (direct calculateBounds call and
  authored branch order); DespawnAllActiveMonsterGroups0048C9F0 is
  exact via the inlined GetAt bounds-check idiom.
- MonsterActionTargetsEnemies rewritten as the switch the retail
  sub/je chain lowers from.
- ScrollTextBoxTo rebuilt around the dormant dialogue-input offset
  (line_count added into the count when the editor is open on the
  cursor's box); W8_TEXT_BOX_VISIBLE_LINE_COUNT is now the real
  macro expression retail inlines, moving ScrollTextBoxTo,
  ScrollTextBoxUp, CurrentDialogueLineHasContent and
  GetTextBoxVisibleLineCount callers to semantic equivalence.
- CopyWorldItemInstance rewritten as the per-field copy retail emits.
- GetFactionDispositionToward reordered to retail's party-first
  branch; residual is VC6 same-TU inlining of GetFactionDisposition.
- SubMenuButtonUseItem expanded to the full combat teardown/open
  flow; residual is the row-array end address.
- ComputeLevelUpSpellPointAward uses retail's level <= 1 predicate.
- HasEventCharacter remaps before mask computation as retail does.
- W8CameraShakeEffect::value_08 retyped to float; new float
  constants 81.25f/32.5f at 0x005ECE58/0x005ECE5C.

+++ include/wiz8/engine_code/GrCycle.h
@@ -52,7 +52,7 @@ void DestroyLightVector(W8GrowableVector<stLight*>* vector); /* 0x004A8C50 */
    Trigger sets bit 4 on its own to reverse the shake. */
 class W8CameraShakeEffect {
 public:
-    W8CameraShakeEffect(float duration, char preset, float intensity, int value_08,
+    W8CameraShakeEffect(float duration, char preset, float intensity, float value_08,
                         const srVector3T<float>* position); /* 0x004ADED0 */
     W8CameraShakeEffect(const W8CameraShakeEffect& other);  /* 0x004AE000 */
     /* Per-frame evaluation: answers whether the effect is still active and
@@ -61,7 +61,7 @@ public:
 
     unsigned int flags_00;         /* 0x00 */
     float intensity_04;            /* 0x04 */
-    int value_08;                  /* 0x08 */
+    float value_08;                /* 0x08: distance cap for bit-2 effects */
     srVector3T<float> position_0c; /* 0x0c */
     W8GameTimer timer_18;          /* 0x18 */
     /* The key 0x004AE170 matches an animation event against. */
@@ -79,7 +79,7 @@ extern W8GrowableVector<W8CameraShakeEffect*>* g_shake_effects_0065be2c;
 extern W8GameTimer* g_shake_timer_0065be30;
 
 W8CameraShakeEffect* CreateCameraShakeEffect004AE080(float duration, char preset, float intensity,
-                                                     int value_08,
+                                                     float value_08,
                                                      const srVector3T<float>* position);
 /* Fire every effect in one cycle's vector whose key matches, moving it onto the
    live list and restarting its timer. */
